обработка статуса в апи регистрации
