--- src/api/Auth/index.ts
@@ -1,2 +1,0 @@
-export * from './Auth.api';
-export { default } from './Auth.api';
