+++ src/api/SignUp/index.ts
@@ -1,0 +1,2 @@
+export * from './SignUp.api';
+export { default } from './SignUp.api';
