+++ src/pages/SignUp/SignUp.view.tsx
@@ -25,7 +25,7 @@ const SignUp:FC = () => {
                     onFinishFailed={onFinishFailed}
                     layout='vertical'>
                         {fieldSet.map(set =>
-                            <Form.Item name={set.name} rules={set.rules}>
+                            <Form.Item key={set.name} name={set.name} rules={set.rules}>
                                 <Input type={set.type ?? ''} placeholder={set.placeholder} />
                             </Form.Item>
                         )}
